adapts decoration descriptions reader and notification dispatcher
now the inheritance hierarchy of stereotypes is also covered when
retrieving decoration descriptions.
same goes with activation conditions (if the child node does not have
one, then the parent is checked for activation code and so on)

Decoration notification dispatcher now considers the previous decoration
status in different scenarios in order to send minimal amount of
notifications as possible to the decorators. The decoration descriptions
are now tracked by the dispatcher.