adds decoration descriptions resource listener
when a profile is applied/loaded the resource change listener will track
workspace changes to determine if descriptions were adapted in editor.
When a resource change is detected the decoration notification
dispatcher takes measures to update visualizations or remove them
completely if there were validation problems. 

---
Also removing system error output in case of executing the unloading of
the profile application command when the editor part is closing.